+++ NEWS.md
@@ -1,3 +1,8 @@
+# naomi 1.0.5
+
+* Add `hintr_calibrate()` which takes output and calibration options and runs
+  calibration and returns in hintr output format
+
 # naomi 1.0.4
 
 * Add `get_calibration_options()` to retrieve options for calibration